% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multsampsize-package.R
\docType{package}
\name{multsampsize-package}
\alias{multsampsize}
\alias{multsampsize-package}
\title{multsampsize: Sample size calculations for clinical prediction models developing using multinomial logistic regression}
\description{
Sample size calculations for clinical prediction models developing using multinomial logistic regression.
}
\author{
\strong{Maintainer}: Alexander Pate \email{alexander.pate@manchester.ac.uk}

}
\keyword{internal}
