% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multsampsize-package.R
\docType{package}
\name{multsampsize-package}
\alias{multsampsize}
\alias{multsampsize-package}
\title{multsampsize: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
